pyproject.toml
[project]
name = "pyswitch"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Molly Draven", email = "drazi@duck.com"},
]
dependencies = [
    "sentry-sdk>=2.12.0",
    "cryptography>=43.0.0",
    "loguru>=0.7.2",
    "python-dotenv>=1.0.1",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false

[tool.pdm.dev-dependencies]
dev = [
    "flake8>=7.1.1",
    "black>=24.8.0",
    "pytest>=8.3.2",
    "pytest-cov>=5.0.0",
]
